requirements.txt
torch == 1.5.0
torchvision == 0.6.0
scikit-learn == 0.23.1
numpy == 1.18.5
pandas == 0.20.3
Pillow == 7.1.2
tensorflow == 2.2.0